Move background_processing
