+++ moveit_core/CMakeLists.txt
@@ -56,7 +56,6 @@ find_package(pluginlib REQUIRED)
 set(VERSION_FILE_PATH ${CMAKE_BINARY_DIR}/include)
 
 set(THIS_PACKAGE_INCLUDE_DIRS
-    background_processing/include
     exceptions/include
     collision_detection/include
     collision_detection_fcl/include
@@ -87,7 +86,6 @@ set(THIS_PACKAGE_INCLUDE_DIRS
 )
 
 set(THIS_PACKAGE_LIBRARIES
-    moveit_background_processing
     moveit_butterworth_filter
     moveit_collision_distance_field
     moveit_collision_detection
@@ -157,7 +155,6 @@ message(STATUS " *** Building MoveIt ${MOVEIT_VERSION} ***")
 configure_file("version/version.h.in" "${VERSION_FILE_PATH}/moveit/version.h")
 install(FILES "${VERSION_FILE_PATH}/moveit/version.h" DESTINATION include/moveit)
 
-add_subdirectory(background_processing)
 add_subdirectory(collision_distance_field)
 add_subdirectory(constraint_samplers)
 add_subdirectory(controller_manager)
