--- moveit_core/background_processing/CMakeLists.txt
@@ -1,12 +1,0 @@
-set(MOVEIT_LIB_NAME moveit_background_processing)
-
-add_library(${MOVEIT_LIB_NAME} SHARED src/background_processing.cpp)
-set_target_properties(${MOVEIT_LIB_NAME} PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
-ament_target_dependencies(${MOVEIT_LIB_NAME}
-  Boost
-  rclcpp
-  urdfdom
-  urdfdom_headers
-)
-
-install(DIRECTORY include/ DESTINATION include)
